packages/main/pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.6",
    "pywebview>=5.4",
]

[dependency-groups]
dev = [
    "nuitka>=2.7.10",
]
